fix: close all GUIs when switching forces, not just main window
If the main window wasn't open when switching forces, the individual todo windows could be left open and cause a crash when a todo was completed in a different surface than the one it was created in. This mitigates that by closing all windows when switching forces, not just the main one.